+++ projects/js-packages/components/components/zendesk-chat/index.tsx
@@ -1,8 +1,16 @@
-import { useEffect } from 'react';
+import { useEffect, useCallback } from 'react';
 import { chatKey } from './constants';
 import type { ZendeskChatType } from './types';
 
-export const ZendeskChat: ZendeskChatType = () => {
+export const ZendeskChat: ZendeskChatType = ( { jwt_token } ) => {
+	const authenticateUser = useCallback( () => {
+		if ( typeof window !== 'undefined' && typeof window.zE === 'function' ) {
+			window.zE( 'messenger', 'loginUser', function ( callback ) {
+				callback( jwt_token );
+			} );
+		}
+	}, [ jwt_token ] );
+
 	useEffect( () => {
 		const script = document.createElement( 'script' );
 		const container = document.getElementById( 'zendesk-chat-container' );
@@ -11,10 +19,14 @@ export const ZendeskChat: ZendeskChatType = () => {
 		script.type = 'text/javascript';
 		script.id = 'ze-snippet';
 
+		script.onload = () => {
+			authenticateUser();
+		};
+
 		if ( container ) {
 			container.appendChild( script );
 		}
-	}, [] );
+	}, [ authenticateUser ] );
 
 	return <div data-testid="zendesk-chat-container" id="zendesk-chat-container" />;
 };