+++ projects/js-packages/components/tools/jp-redirect/index.ts
@@ -25,7 +25,7 @@ export default function getRedirectUrl( source: string, args: GetRedirectUrlArgs
 
 	let calypsoEnv;
 	if ( typeof window !== 'undefined' ) {
-		calypsoEnv = window.Initial_State?.calypsoEnv;
+		calypsoEnv = window?.JP_CONNECTION_INITIAL_STATE?.calypsoEnv;
 	}
 
 	if ( source.search( 'https://' ) === 0 ) {
@@ -38,9 +38,9 @@ export default function getRedirectUrl( source: string, args: GetRedirectUrlArgs
 		queryVars.source = encodeURIComponent( source );
 	}
 
-	Object.keys( args ).map( argName => {
+	for ( const argName in args ) {
 		queryVars[ argName ] = encodeURIComponent( args[ argName ] );
-	} );
+	}
 
 	if (
 		! Object.keys( queryVars ).includes( 'site' ) &&