+++ projects/packages/status/src/class-host.php
@@ -93,4 +93,30 @@ class Host {
 		$domains[] = 'calypso.localhost';
 		return $domains;
 	}
+
+	/**
+	 * Return Calypso environment value; used for developing Jetpack and pairing
+	 * it with different Calypso environments, such as localhost.
+	 *
+	 * @since 1.18.0
+	 *
+	 * @return string Calypso environment
+	 */
+	public function get_calypso_env() {
+		// phpcs:disable WordPress.Security.NonceVerification.Recommended -- Nonce is not required; only used for changing environments.
+		if ( isset( $_GET['calypso_env'] ) ) {
+			return sanitize_key( $_GET['calypso_env'] );
+		}
+		// phpcs:enable WordPress.Security.NonceVerification.Recommended
+
+		if ( getenv( 'CALYPSO_ENV' ) ) {
+			return sanitize_key( getenv( 'CALYPSO_ENV' ) );
+		}
+
+		if ( defined( 'CALYPSO_ENV' ) && CALYPSO_ENV ) {
+			return sanitize_key( CALYPSO_ENV );
+		}
+
+		return '';
+	}
 }