+++ projects/plugins/debug-helper/modules/class-waf-helper.php
@@ -182,7 +182,7 @@ class Waf_Helper {
 		<hr>
 
 		<h2>Status</h2>
-		<p>Environment is supported: <code><?php echo Waf_Runner::is_supported_environment() ? 'true' : 'false'; ?></code></p>
+		<p>WAF is supported: <code><?php echo Waf_Runner::is_supported_environment() ? 'true' : 'false'; ?></code></p>
 		<p>Firewall status: <code><?php echo Waf_Runner::is_enabled() ? 'enabled' : 'disabled'; ?></code></p>
 
 		<hr>