+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/extensions/ai-assistant/index.ts
@@ -3,7 +3,6 @@
  */
 import { getBlockType } from '@wordpress/blocks';
 import { select } from '@wordpress/data';
-import { store as editPostStore } from '@wordpress/edit-post';
 import { addFilter } from '@wordpress/hooks';
 /*
  * Internal dependencies
@@ -18,9 +17,9 @@ import { isUserConnected } from '../../lib/connection';
 export const AI_ASSISTANT_SUPPORT_NAME = 'ai-assistant-support';
 
 // List of blocks that can be extended.
-export const EXTENDED_BLOCKS = [ 'core/paragraph', 'core/heading', 'core/list-item' ] as const;
+export const EXTENDED_BLOCKS = [ 'core/paragraph', 'core/heading', 'core/list' ] as const;
 
-type ExtendedBlock = ( typeof EXTENDED_BLOCKS )[ number ];
+export type ExtendedBlockProp = ( typeof EXTENDED_BLOCKS )[ number ];
 
 type BlockSettingsProps = {
 	supports: {
@@ -32,8 +31,6 @@ type BlockSettingsProps = {
 
 export const isAiAssistantSupportExtensionEnabled =
 	window?.Jetpack_Editor_Initial_State.available_blocks?.[ AI_ASSISTANT_SUPPORT_NAME ];
-
-const siteRequiresUpgrade = AI_Assistant_Initial_State.requireUpgrade;
 
 /**
  * Check if it is possible to extend the block.
@@ -57,19 +54,18 @@ export function isPossibleToExtendBlock(): boolean {
 		return false;
 	}
 
-	// Do not extend the block if the site requires an upgrade.
-	if ( siteRequiresUpgrade ) {
-		return false;
-	}
-
 	// Do not extend if there is an error getting the feature.
 	if ( AI_Assistant_Initial_State.errorCode ) {
 		return false;
 	}
 
-	// Do not extend if the AI Assistant block is hidden
-	const { getHiddenBlockTypes } = select( editPostStore );
-	const hiddenBlocks = getHiddenBlockTypes();
+	/*
+	 * Do not extend if the AI Assistant block is hidden
+	 * ToDo: the `editPostStore` is undefined for P2 sites.
+	 * Let's find a way to check if the block is hidden.
+	 */
+	const { getHiddenBlockTypes } = select( 'core/edit-post' ) || {};
+	const hiddenBlocks = getHiddenBlockTypes?.() || []; // It will extend the block if the function is undefined.
 	if ( hiddenBlocks.includes( blockName ) ) {
 		return false;
 	}
@@ -81,12 +77,12 @@ export function isPossibleToExtendBlock(): boolean {
  * Add jetpack/ai support to the extended blocks.
  *
  * @param {BlockSettingsProps} settings - Block settings.
- * @param {ExtendedBlock} name          - Block name.
+ * @param {ExtendedBlockProp} name          - Block name.
  * @returns {BlockSettingsProps}          Block settings.
  */
 function addJetpackAISupport(
 	settings: BlockSettingsProps,
-	name: ExtendedBlock
+	name: ExtendedBlockProp
 ): BlockSettingsProps {
 	// Only extend the blocks in the list.
 	if ( ! EXTENDED_BLOCKS.includes( name ) ) {