+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/lib/suggestions/index.js
@@ -1,12 +1,9 @@
 /**
  * External dependencies
  */
+import { requestJwt } from '@automattic/jetpack-ai-client';
 import { fetchEventSource } from '@microsoft/fetch-event-source';
 import debugFactory from 'debug';
-/**
- * Internal dependencies
- */
-import { requestToken } from './token';
 
 const debug = debugFactory( 'jetpack-ai-assistant' );
 
@@ -42,40 +39,23 @@ export async function askJetpack( question ) {
  * @param {object} options                    - Options
  * @param {number} options.postId             - The post where this completion is being requested, if available
  * @param {boolean} options.fromCache         - Get a cached response. False by default.
- * @param {boolean} options.requireUpgrade    - If the site requires an upgrade to use the feature
+ * @param {boolean} options.useGpt4           - If the request should use GPT-4
  * @returns {Promise<SuggestionsEventSource>} The event source
  */
 export async function askQuestion(
 	question,
-	{ postId = null, fromCache = false, requireUpgrade }
+	{ postId = null, fromCache = false, useGpt4 = false }
 ) {
-	if ( requireUpgrade ) {
-		/*
-		 * Return an empty event source
-		 * @todo: ideally, business part shouln't be here
-		 */
-		return new SuggestionsEventSource( '' );
-	}
-
-	const { token } = await requestToken();
+	const { token } = await requestJwt();
 
 	const url = new URL( 'https://public-api.wordpress.com/wpcom/v2/jetpack-ai-query' );
-	if ( Array.isArray( question ) ) {
-		url.searchParams.append( 'messages', JSON.stringify( question ) );
-	} else {
-		url.searchParams.append( 'question', question );
-	}
-	url.searchParams.append( 'token', token );
+	const options = { postId, useGpt4 };
 
 	if ( fromCache ) {
 		url.searchParams.append( 'stream_cache', 'true' );
 	}
 
-	if ( postId ) {
-		url.searchParams.append( 'post_id', postId );
-	}
-
-	return new SuggestionsEventSource( url.toString() );
+	return new SuggestionsEventSource( { url: url.toString(), question, token, options } );
 }
 
 /**
@@ -93,19 +73,36 @@ export async function askQuestion(
  * @fires error_network - The EventSource connection to the server returned some error
  */
 export class SuggestionsEventSource extends EventTarget {
-	constructor( url ) {
+	constructor( data ) {
 		super();
 		this.fullMessage = '';
 		this.isPromptClear = false;
 		// The AbortController is used to close the fetchEventSource connection
 		this.controller = new AbortController();
-		this.initEventSource( url );
+		this.initEventSource( data );
 	}
 
-	async initEventSource( url ) {
+	async initEventSource( { url, question, token, options = {} } = {} ) {
 		const self = this;
+		const bodyData = { post_id: options?.postId };
+
+		if ( Array.isArray( question ) ) {
+			bodyData.messages = question;
+		} else {
+			bodyData.question = question;
+		}
+
+		if ( options?.useGpt4 ) {
+			bodyData.feature = 'ai-assistant-experimental';
+		}
 
 		this.source = await fetchEventSource( url.toString(), {
+			method: 'POST',
+			headers: {
+				'Content-type': 'application/json',
+				Authorization: 'Bearer ' + token,
+			},
+			body: JSON.stringify( bodyData ),
 			onclose() {
 				debug( 'Stream closed unexpectedly' );
 			},