+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/lib/utils/block-content.ts
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { store as blockEditorStore } from '@wordpress/block-editor';
 import { getBlockContent } from '@wordpress/blocks';
 import { serialize } from '@wordpress/blocks';
 import { select } from '@wordpress/data';
@@ -9,8 +8,6 @@ import TurndownService from 'turndown';
 
 // Turndown instance
 const turndownService = new TurndownService();
-
-const HTML_JOIN_CHARACTERS = '<br />';
 
 /**
  * Returns partial content from the beginning of the post
@@ -52,49 +49,20 @@ export function getContentFromBlocks(): string {
 	return turndownService.turndown( serialize( blocks ) );
 }
 
-type GetTextContentFromBlocksProps = {
-	count: number;
-	clientIds: string[];
-	content: string;
-};
-
-/**
- * Returns the text content from all selected blocks.
- *
- * @returns {GetTextContentFromBlocksProps} The text content.
- */
-export function getTextContentFromBlocks(): GetTextContentFromBlocksProps {
-	const clientIds = select( blockEditorStore ).getSelectedBlockClientIds();
-	const defaultContent = {
-		count: 0,
-		clientIds: [],
-		content: '',
-	};
-
-	if ( ! clientIds?.length ) {
-		return defaultContent;
+export function getTextContentFromInnerBlocks( clientId: string ) {
+	const block = select( 'core/block-editor' ).getBlock( clientId );
+	if ( ! block?.innerBlocks?.length ) {
+		return '';
 	}
 
-	const blocks = select( blockEditorStore ).getBlocksByClientId( clientIds );
-	if ( ! blocks?.length ) {
-		return defaultContent;
-	}
-
-	return {
-		count: blocks.length,
-		clientIds,
-		content: blocks
-			.map( block => getBlockTextContent( block.clientId ) )
-			.join( HTML_JOIN_CHARACTERS ),
-	};
+	return block.innerBlocks
+		.filter( blq => blq != null ) // Safeguard against null or undefined blocks
+		.map( blq => getBlockTextContent( blq.clientId ) )
+		.join( '\n\n' );
 }
 
 /**
- * Return the block content from the given block clientId.
- *
- * It will try to get the content from the block `content` attribute.
- * Otherwise, it will try to get the content
- * by using the `getBlockContent` function.
+ * Return the block content from the given block clientId using the `getBlockContent` function.
  *
  * @param {string} clientId   - The block clientId.
  * @returns {string}            The block content.
@@ -104,7 +72,7 @@ export function getBlockTextContent( clientId: string ): string {
 		return '';
 	}
 
-	const editor = select( blockEditorStore );
+	const editor = select( 'core/block-editor' );
 	const block = editor.getBlock( clientId );
 
 	/*
@@ -115,10 +83,33 @@ export function getBlockTextContent( clientId: string ): string {
 		return '';
 	}
 
-	// Attempt to pick the content from the block `content` attribute.
-	if ( block?.attributes?.content ) {
-		return block.attributes.content;
+	return getBlockContent( block );
+}
+
+/**
+ * Extract raw text from HTML content
+ *
+ * @param {string} htmlString - The HTML content.
+ * @returns {string}            The raw text.
+ */
+export function getRawTextFromHTML( htmlString: string ): string {
+	// Removes all continuous whitespace from the start to check if the string is empty
+	if ( ! htmlString?.replace( /\s+/, '' ).length ) {
+		return '';
 	}
 
-	return getBlockContent( block );
+	const tempDomContainer = document.createElement( 'div' );
+	tempDomContainer.innerHTML = htmlString;
+
+	const { textContent, innerText } = tempDomContainer;
+
+	if ( !! textContent && ! textContent.replace( /\s+/, '' ).length ) {
+		return '';
+	}
+
+	if ( !! innerText && ! innerText.replace( /\s+/, '' ).length ) {
+		return '';
+	}
+
+	return tempDomContainer.textContent || tempDomContainer.innerText || '';
 }