+++ projects/packages/assets/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.18.7] - 2023-07-11
+### Changed
+- Updated package dependencies. [#31785]
+
+## [1.18.6] - 2023-07-05
+### Changed
+- Updated package dependencies. [#31659]
+
 ## [1.18.5] - 2023-06-21
 ### Changed
 - Updated package dependencies. [#31468]
@@ -337,6 +345,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Statically access asset tools
 
+[1.18.7]: https://github.com/Automattic/jetpack-assets/compare/v1.18.6...v1.18.7
+[1.18.6]: https://github.com/Automattic/jetpack-assets/compare/v1.18.5...v1.18.6
 [1.18.5]: https://github.com/Automattic/jetpack-assets/compare/v1.18.4...v1.18.5
 [1.18.4]: https://github.com/Automattic/jetpack-assets/compare/v1.18.3...v1.18.4
 [1.18.3]: https://github.com/Automattic/jetpack-assets/compare/v1.18.2...v1.18.3